Added check for scene_numbering table existence when upgrading main database to v16

+++ sickrage/core/databases/main/migrations/versions/016_Merge_Scene_Numbering_Table_With_TVEpisodes_Table.py
@@ -16,17 +16,17 @@ down_revision = '15'
 def upgrade():
     conn = op.get_bind()
     meta = sa.MetaData(bind=conn)
-    tv_episodes = sa.Table('tv_episodes', meta, autoload=True)
-    scene_numbering = sa.Table('scene_numbering', meta, autoload=True)
 
-    with op.get_context().begin_transaction():
-        for row in conn.execute(scene_numbering.select()):
-            conn.execute('UPDATE tv_episodes SET scene_season = {} WHERE tv_episodes.showid = {}, tv_episodes.season = {}, tv_episodes.episode = {}'
-                         .format(row.scene_season, row.indexer_id, row.season, row.episode))
-            conn.execute('UPDATE tv_episodes SET scene_episode = {} WHERE tv_episodes.showid = {}, tv_episodes.season = {}, tv_episodes.episode = {}'
-                         .format(row.scene_episode, row.indexer_id, row.season, row.episode))
+    if conn.engine.dialect.has_table(conn.engine, 'scene_numbering'):
+        scene_numbering = sa.Table('scene_numbering', meta, autoload=True)
+        with op.get_context().begin_transaction():
+            for row in conn.execute(scene_numbering.select()):
+                conn.execute('UPDATE tv_episodes SET scene_season = {} WHERE tv_episodes.showid = {}, tv_episodes.season = {}, tv_episodes.episode = {}'
+                             .format(row.scene_season, row.indexer_id, row.season, row.episode))
+                conn.execute('UPDATE tv_episodes SET scene_episode = {} WHERE tv_episodes.showid = {}, tv_episodes.season = {}, tv_episodes.episode = {}'
+                             .format(row.scene_episode, row.indexer_id, row.season, row.episode))
 
-    op.drop_table('scene_numbering')
+        op.drop_table('scene_numbering')
 
 
 def downgrade():
